Renamed the trackerUserPermission , UserDashboard and typesense folder in the modules directory and updated all import paths where the folder is used.
Renamed the trackerUserPermission , UserDashboard and typesense folder in the modules directory and updated all import paths where the folder is used.
